extracting util code to new submodule
