+++ src/py/crankshaft/crankshaft/clustering/moran.py
@@ -5,9 +5,11 @@ Moran's I geostatistics (global clustering & outliers presence)
 # TODO: Fill in local neighbors which have null/NoneType values with the
 #       average of the their neighborhood
 
-import numpy as np
 import pysal as ps
 import plpy
+
+# crankshaft module
+import crankshaft.pysal_utils as pu
 
 # High level interface ---------------------------------------
 
@@ -25,7 +27,7 @@ def moran(subquery, attr_name,
              "subquery": subquery,
              "num_ngbrs": num_ngbrs}
 
-    query = construct_neighbor_query(w_type, qvals)
+    query = pu.construct_neighbor_query(w_type, qvals)
 
     plpy.notice('** Query: %s' % query)
 
@@ -33,22 +35,23 @@ def moran(subquery, attr_name,
         result = plpy.execute(query)
         # if there are no neighbors, exit
         if len(result) == 0:
-            return empty_zipped_array(2)
+            return pu.empty_zipped_array(2)
         plpy.notice('** Query returned with %d rows' % len(result))
     except plpy.SPIError:
         plpy.error('Error: areas of interest query failed, check input parameters')
         plpy.notice('** Query failed: "%s"' % query)
         plpy.notice('** Error: %s' % plpy.SPIError)
-        return empty_zipped_array(2)
+        return pu.empty_zipped_array(2)
 
     ## collect attributes
-    attr_vals = get_attributes(result)
+    attr_vals = pu.get_attributes(result)
 
     ## calculate weights
-    weight = get_weight(result, w_type, num_ngbrs)
+    weight = pu.get_weight(result, w_type, num_ngbrs)
 
     ## calculate moran global
-    moran_global = ps.esda.moran.Moran(attr_vals, weight, permutations=permutations)
+    moran_global = ps.esda.moran.Moran(attr_vals, weight,
+                                       permutations=permutations)
 
     return zip([moran_global.I], [moran_global.EI])
 
@@ -68,20 +71,20 @@ def moran_local(subquery, attr,
              "subquery": subquery,
              "num_ngbrs": num_ngbrs}
 
-    query = construct_neighbor_query(w_type, qvals)
+    query = pu.construct_neighbor_query(w_type, qvals)
 
     try:
         result = plpy.execute(query)
         # if there are no neighbors, exit
         if len(result) == 0:
-            return empty_zipped_array(5)
+            return pu.empty_zipped_array(5)
     except plpy.SPIError:
         plpy.error('Error: areas of interest query failed, check input parameters')
         plpy.notice('** Query failed: "%s"' % query)
-        return empty_zipped_array(5)
+        return pu.empty_zipped_array(5)
 
-    attr_vals = get_attributes(result)
-    weight = get_weight(result, w_type)
+    attr_vals = pu.get_attributes(result)
+    weight = pu.get_weight(result, w_type, num_ngbrs)
 
     # calculate LISA values
     lisa = ps.esda.moran.Moran_Local(attr_vals, weight,
@@ -90,7 +93,6 @@ def moran_local(subquery, attr,
     # find quadrants for each geometry
     quads = quad_position(lisa.q)
 
-    plpy.notice('** Finished calculations')
     return zip(lisa.Is, quads, lisa.p_sim, weight.id_order, lisa.y)
 
 def moran_rate(subquery, numerator, denominator,
@@ -106,7 +108,7 @@ def moran_rate(subquery, numerator, denominator,
              "subquery": subquery,
              "num_ngbrs": num_ngbrs}
 
-    query = construct_neighbor_query(w_type, qvals)
+    query = pu.construct_neighbor_query(w_type, qvals)
 
     plpy.notice('** Query: %s' % query)
 
@@ -114,19 +116,19 @@ def moran_rate(subquery, numerator, denominator,
         result = plpy.execute(query)
         # if there are no neighbors, exit
         if len(result) == 0:
-            return empty_zipped_array(2)
+            return pu.empty_zipped_array(2)
         plpy.notice('** Query returned with %d rows' % len(result))
     except plpy.SPIError:
         plpy.error('Error: areas of interest query failed, check input parameters')
         plpy.notice('** Query failed: "%s"' % query)
         plpy.notice('** Error: %s' % plpy.SPIError)
-        return empty_zipped_array(2)
+        return pu.empty_zipped_array(2)
 
     ## collect attributes
-    numer = get_attributes(result, 1)
-    denom = get_attributes(result, 2)
+    numer = pu.get_attributes(result, 1)
+    denom = pu.get_attributes(result, 2)
 
-    weight = get_weight(result, w_type, num_ngbrs)
+    weight = pu.get_weight(result, w_type, num_ngbrs)
 
     ## calculate moran global rate
     lisa_rate = ps.esda.moran.Moran_Rate(numer, denom, weight,
@@ -143,7 +145,7 @@ def moran_local_rate(subquery, numerator, denominator,
     # geometries with values that are null are ignored
     # resulting in a collection of not as near neighbors
 
-    query = construct_neighbor_query(w_type,
+    query = pu.construct_neighbor_query(w_type,
                                      {"id_col": id_col,
                                       "numerator": numerator,
                                       "denominator": denominator,
@@ -155,18 +157,18 @@ def moran_local_rate(subquery, numerator, denominator,
         result = plpy.execute(query)
         # if there are no neighbors, exit
         if len(result) == 0:
-            return empty_zipped_array(5)
+            return pu.empty_zipped_array(5)
     except plpy.SPIError:
         plpy.error('Error: areas of interest query failed, check input parameters')
         plpy.notice('** Query failed: "%s"' % query)
         plpy.notice('** Error: %s' % plpy.SPIError)
-        return empty_zipped_array(5)
+        return pu.empty_zipped_array(5)
 
     ## collect attributes
-    numer = get_attributes(result, 1)
-    denom = get_attributes(result, 2)
+    numer = pu.get_attributes(result, 1)
+    denom = pu.get_attributes(result, 2)
 
-    weight = get_weight(result, w_type, num_ngbrs)
+    weight = pu.get_weight(result, w_type, num_ngbrs)
 
     # calculate LISA values
     lisa = ps.esda.moran.Moran_Local_Rate(numer, denom, weight,
@@ -191,25 +193,25 @@ def moran_local_bv(subquery, attr1, attr2,
              "geom_col": geom_col,
              "id_col": id_col}
 
-    query = construct_neighbor_query(w_type, qvals)
+    query = pu.construct_neighbor_query(w_type, qvals)
 
     try:
         result = plpy.execute(query)
         # if there are no neighbors, exit
         if len(result) == 0:
-            return empty_zipped_array(4)
+            return pu.empty_zipped_array(4)
     except plpy.SPIError:
         plpy.error("Error: areas of interest query failed, " \
                    "check input parameters")
         plpy.notice('** Query failed: "%s"' % query)
-        return empty_zipped_array(4)
+        return pu.empty_zipped_array(4)
 
     ## collect attributes
-    attr1_vals = get_attributes(result, 1)
-    attr2_vals = get_attributes(result, 2)
+    attr1_vals = pu.get_attributes(result, 1)
+    attr2_vals = pu.get_attributes(result, 2)
 
     # create weights
-    weight = get_weight(result, w_type, num_ngbrs)
+    weight = pu.get_weight(result, w_type, num_ngbrs)
 
     # calculate LISA values
     lisa = ps.esda.moran.Moran_Local_BV(attr1_vals, attr2_vals, weight,
@@ -246,138 +248,6 @@ def map_quads(coord):
     else:
         return None
 
-def query_attr_select(params):
-    """
-        Create portion of SELECT statement for attributes inolved in query.
-        @param params: dict of information used in query (column names,
-                       table name, etc.)
-    """
-
-    attrs = [k for k in params
-             if k not in ('id_col', 'geom_col', 'subquery', 'num_ngbrs')]
-
-    template = "i.\"{%(col)s}\"::numeric As attr%(alias_num)s, "
-
-    attr_string = ""
-
-    for idx, val in enumerate(sorted(attrs)):
-        attr_string += template % {"col": val, "alias_num": idx + 1}
-
-    return attr_string
-
-def query_attr_where(params):
-    """
-        Create portion of WHERE clauses for weeding out NULL-valued geometries
-    """
-    attrs = sorted([k for k in params
-                    if k not in ('id_col', 'geom_col', 'subquery', 'num_ngbrs')])
-
-    attr_string = []
-
-    for attr in attrs:
-        attr_string.append("idx_replace.\"{%s}\" IS NOT NULL" % attr)
-
-    if len(attrs) == 2:
-        attr_string.append("idx_replace.\"{%s}\" <> 0" % attrs[1])
-
-    out = " AND ".join(attr_string)
-
-    return out
-
-def knn(params):
-    """SQL query for k-nearest neighbors.
-        @param vars: dict of values to fill template
-    """
-
-    attr_select = query_attr_select(params)
-    attr_where = query_attr_where(params)
-
-    replacements = {"attr_select": attr_select,
-                    "attr_where_i": attr_where.replace("idx_replace", "i"),
-                    "attr_where_j": attr_where.replace("idx_replace", "j")}
-
-    query = "SELECT " \
-                "i.\"{id_col}\" As id, " \
-                "%(attr_select)s" \
-                "(SELECT ARRAY(SELECT j.\"{id_col}\" " \
-                              "FROM ({subquery}) As j " \
-                              "WHERE %(attr_where_j)s AND " \
-                                    "i.\"{id_col}\" <> j.\"{id_col}\" " \
-                              "ORDER BY j.\"{geom_col}\" <-> i.\"{geom_col}\" ASC " \
-                              "LIMIT {num_ngbrs}) " \
-                ") As neighbors " \
-            "FROM ({subquery}) As i " \
-            "WHERE " \
-                "%(attr_where_i)s " \
-            "ORDER BY i.\"{id_col}\" ASC;" % replacements
-
-    return query.format(**params)
-
-## SQL query for finding queens neighbors (all contiguous polygons)
-def queen(params):
-    """SQL query for queen neighbors.
-        @param params dict: information to fill query
-    """
-    attr_select = query_attr_select(params)
-    attr_where = query_attr_where(params)
-
-    replacements = {"attr_select": attr_select,
-                    "attr_where_i": attr_where.replace("idx_replace", "i"),
-                    "attr_where_j": attr_where.replace("idx_replace", "j")}
-
-    query = "SELECT " \
-                "i.\"{id_col}\" As id, " \
-                "%(attr_select)s" \
-                "(SELECT ARRAY(SELECT j.\"{id_col}\" " \
-                 "FROM ({subquery}) As j " \
-                 "WHERE ST_Touches(i.\"{geom_col}\", j.\"{geom_col}\") AND " \
-                 "%(attr_where_j)s)" \
-                ") As neighbors " \
-            "FROM ({subquery}) As i " \
-            "WHERE " \
-                "%(attr_where_i)s " \
-            "ORDER BY i.\"{id_col}\" ASC;" % replacements
-
-    return query.format(**params)
-
-## to add more weight methods open a ticket or pull request
-
-def construct_neighbor_query(w_type, query_vals):
-    """Return requested query.
-        @param w_type text: type of neighbors to calculate ('knn' or 'queen')
-        @param query_vals dict: values used to construct the query
-    """
-
-    if w_type == 'knn':
-        return knn(query_vals)
-    else:
-        return queen(query_vals)
-
-def get_attributes(query_res, attr_num=1):
-    """
-        @param query_res: query results with attributes and neighbors
-        @param attr_num: attribute number (1, 2, ...)
-    """
-    return np.array([x['attr' + str(attr_num)] for x in query_res], dtype=np.float)
-
-## Build weight object
-def get_weight(query_res, w_type='queen', num_ngbrs=5):
-    """
-        Construct PySAL weight from return value of query
-        @param query_res: query results with attributes and neighbors
-    """
-    if w_type == 'knn':
-        row_normed_weights = [1.0 / float(num_ngbrs)] * num_ngbrs
-        weights = {x['id']: row_normed_weights for x in query_res}
-    else:
-        weights = {x['id']: [1.0 / len(x['neighbors'])] * len(x['neighbors'])
-                            if len(x['neighbors']) > 0
-                            else [] for x in query_res}
-
-    neighbors = {x['id']: x['neighbors'] for x in query_res}
-
-    return ps.W(neighbors, weights)
-
 def quad_position(quads):
     """
         Produce Moran's I classification based of n
@@ -385,20 +255,6 @@ def quad_position(quads):
         @param quads ndarray: an array of quads classified by
           1-4 (PySAL default)
         Output:
-        @param ndarray: an array of quads classied by 'HH', 'LL', etc.
+        @param list: an array of quads classied by 'HH', 'LL', etc.
     """
-
-    lisa_sig = np.array([map_quads(q) for q in quads])
-
-    return lisa_sig
-
-def return_empty_zipped_array(num_nones):
-    """
-        prepare return values for cases of empty weights objects (no neighbors)
-        Input:
-        @param num_nones int: number of columns (e.g., 4)
-        Output:
-        [(None, None, None, None)]
-    """
-
-    return [tuple([None] * num_nones)]
+    return [map_quads(q) for q in quads]
